OpenCOLLADA/CMakeLists.txt: Build type, GNUInstallDirs, and version fixes
Use CMAKE_BUILD_TYPE which can be overrided with custom distro types

Adhere to GNU filesystem layout conventions. Not all distros and OSes
have directories in the same locations and they define the
GNUInstallDirs variables, so use those variables. Also, this enables
multi lib installations.

Define the version variables to enable SONAME versioning on the
library files in case there is ABI changes in the future.